+++ Src/Circuit/Tree/ClockTreeRRand.cpp
@@ -82,7 +82,7 @@ int ClockTreeRRand::GenerateCkt()
             fanoutList.push_back(RandomInt(minFanout, maxFanout));
 
         currLevelRIndex = 0;
-        for (int j = 0; j < fanoutList.size(); ++ j) {
+        for (size_t j = 0; j < fanoutList.size(); ++ j) {
             k = fanoutList.at(j);
             for (int m = 0; m < k; ++ m) {
                 posNodeIndex = nodeIndex + j;
@@ -105,7 +105,7 @@ int ClockTreeRRand::GenerateCkt()
         fanoutList.push_back(RandomInt(minFanout, maxFanout));
     }
     currLevelRIndex = 0;
-    for (int j = 0; j < fanoutList.size(); ++ j) {
+    for (size_t j = 0; j < fanoutList.size(); ++ j) {
         k = fanoutList.at(j);
         for (int m = 0; m < k; ++ m) {
             posNodeIndex = nodeIndex + j;
